chore: move TokenBucket type to new type file under Algorithm
