+++ src/Domain/Algorithm/types.ts
@@ -1,0 +1,11 @@
+import { Rate } from "../types.ts";
+
+export type TokenBucketState = {
+    tokensCount: number;
+    lastUpdatedAtInMs: number;
+};
+
+export type TokenBucketConfig = {
+    bucketCapacity: number;
+    refillRate: Rate;
+};
